+++ src/App.tsx
@@ -13,8 +13,8 @@ import PrivateLayout from './components/component/PrivateLayout/PrivateLayout';
 
 import SignIn from './components/component/PublicLayout/Authorization/SignIn/SignIn';
 import SignUp from './components/component/PublicLayout/Authorization/SignUp/SignUp';
-import MainPageContainer from './components/component/PrivateLayout/MainPage/MainPageContainer';
-import SettingsContainer from './components/component/PrivateLayout/Settings/SettingsContainer';
+import MainPage from './components/component/PrivateLayout/MainPage/MainPage';
+import SettingsPage from './components/component/PrivateLayout/Settings/SettingsPage';
 import { PageNotFound } from './components/common/PageNotFound/PageNotFound';
 
 
@@ -31,8 +31,8 @@ function App() {
         </Route>
         <Route element={<PrivateProvider />}>
           <Route element={<PrivateLayout />}>
-            <Route path="/mainPage" element={<MainPageContainer />} />
-            <Route path="/settings" element={<SettingsContainer />} />
+            <Route path="/events" element={<MainPage />} />
+            <Route path="/settings" element={<SettingsPage />} />
           </Route>
         </Route>
         <Route path="*" element={<PageNotFound />} />
