+++ src/components/component/PrivateLayout/MainPage/MainPage.tsx
@@ -1,41 +1,26 @@
 import React from 'react';
-import { Grid } from '@mui/material';
 
-import styles from './MainPage.module.scss';
+import PageComponent from '../PageComponent/PageComponent';
+import Main from './Main';
 
-interface IProps {
+interface IState {
   city: string
   description: string
   img: null
 }
 
-const MainPage: React.FC<IProps> = ({ data }) => {
+const data: IState[] = [
+  { city: "Kiev", description: "place", img: null },
+  { city: "Kiev", description: "place", img: null },
+  { city: "Kiev", description: "place", img: null }
+]
 
+const MainPage = () => {
   return (
-    <Grid className={styles.mainPage}>
-      <Grid className={styles.mainPage__filter}>
-        <input placeholder="Filters" />
-      </Grid>
-      <Grid className={styles.mainPage__container}>
-        { data.map(el => (
-          <Grid className={styles.mainPage__content}>
-            <Grid className={styles.mainPage__content_image}>
-              <img src={el.img} alt="image"/>
-            </Grid>
-            <Grid className={styles.mainPage__contentBlock}>
-              <Grid>
-                <Grid className={styles.mainPage__contentBlock_title}>Title</Grid>
-                <Grid className={styles.mainPage__contentBlock_info}>
-                  <p><b>City: {el.city}</b> </p>
-                  <p><b>Description: {el.description}</b> </p>
-                </Grid>
-              </Grid>
-            </Grid>
-          </Grid>
-        ))}
-      </Grid>
-    </Grid>
+    <PageComponent title="Main Page" action>
+      <Main data={data} />
+    </PageComponent>
   )
 }
 
-export default MainPage
+export default MainPage;
